Added damping, fixed explosion with excitation
Added damping dependent from 1 time derivative, implemented both with center and backwards derivative. Center derivative seems to need a very big damping factor to work properly. For backwards something weird happens with beta >= 3.427

Excitation needs to be applied to u(2) and NOT to u(1) as the latter is never updated